Bug 1669966: Don't compare size_t to int r=nbp
The same 'issue' exists in the upstream V8 code this was copied from: https://github.com/v8/v8/blob/c4a7bd869851ecbebc07e83db970c878174f134a/src/utils/vector.h#L49

Differential Revision: https://phabricator.services.mozilla.com/D92982

UltraBlame original commit: bf878ce3529bc093b2467eb3ff7880886a91c610

+++ js/src/irregexp/util/VectorShim.h
@@ -69,7 +69,7 @@ class Vector {
   
   
   int length() const {
-    MOZ_ASSERT(length_ <= std::numeric_limits<int>::max());
+    MOZ_ASSERT(length_ <= static_cast<size_t>(std::numeric_limits<int>::max()));
     return static_cast<int>(length_);
   }
